+++ packages/backend-plugin-api/src/wiring/factories.ts
@@ -17,9 +17,11 @@
 import {
   BackendModuleRegistrationPoints,
   BackendPluginRegistrationPoints,
-  BackendRegistrationPoints,
   BackendFeature,
   ExtensionPoint,
+  InternalBackendFeature,
+  InternalBackendModuleRegistration,
+  InternalBackendPluginRegistration,
 } from './types';
 
 /**
@@ -86,19 +88,62 @@ export interface BackendPluginConfig {
 export function createBackendPlugin<TOptions extends [options?: object] = []>(
   config: BackendPluginConfig | ((...params: TOptions) => BackendPluginConfig),
 ): (...params: TOptions) => BackendFeature {
-  if (typeof config === 'function') {
-    return (...options: TOptions) => {
-      const c = config(...options);
-      return {
-        ...c,
-        id: c.pluginId,
-      };
+  const configCallback = typeof config === 'function' ? config : () => config;
+  return (...options: TOptions): InternalBackendFeature => {
+    const c = configCallback(...options);
+
+    let registrations: InternalBackendPluginRegistration[];
+
+    return {
+      $$type: '@backstage/BackendFeature',
+      version: 'v1',
+      getRegistrations() {
+        if (registrations) {
+          return registrations;
+        }
+        const extensionPoints: InternalBackendPluginRegistration['extensionPoints'] =
+          [];
+        let init: InternalBackendPluginRegistration['init'] | undefined =
+          undefined;
+
+        c.register({
+          registerExtensionPoint(ext, impl) {
+            if (init) {
+              throw new Error(
+                'registerExtensionPoint called after registerInit',
+              );
+            }
+            extensionPoints.push([ext, impl]);
+          },
+          registerInit(regInit) {
+            if (init) {
+              throw new Error('registerInit must only be called once');
+            }
+            init = {
+              deps: regInit.deps,
+              func: regInit.init,
+            };
+          },
+        });
+
+        if (!init) {
+          throw new Error(
+            `registerInit was not called by register in ${c.pluginId}`,
+          );
+        }
+
+        registrations = [
+          {
+            type: 'plugin',
+            pluginId: c.pluginId,
+            extensionPoints,
+            init,
+          },
+        ];
+        return registrations;
+      },
     };
-  }
-  return () => ({
-    ...config,
-    id: config.pluginId,
-  });
+  };
 }
 
 /**
@@ -132,21 +177,50 @@ export interface BackendModuleConfig {
 export function createBackendModule<TOptions extends [options?: object] = []>(
   config: BackendModuleConfig | ((...params: TOptions) => BackendModuleConfig),
 ): (...params: TOptions) => BackendFeature {
-  if (typeof config === 'function') {
-    return (...options: TOptions) => {
-      const c = config(...options);
-      return {
-        id: `${c.pluginId}.${c.moduleId}`,
-        register: c.register,
-      };
+  const configCallback = typeof config === 'function' ? config : () => config;
+  return (...options: TOptions): InternalBackendFeature => {
+    const c = configCallback(...options);
+
+    let registrations: InternalBackendModuleRegistration[];
+
+    return {
+      $$type: '@backstage/BackendFeature',
+      version: 'v1',
+      getRegistrations() {
+        if (registrations) {
+          return registrations;
+        }
+        let init: InternalBackendModuleRegistration['init'] | undefined =
+          undefined;
+
+        c.register({
+          registerInit(regInit) {
+            if (init) {
+              throw new Error('registerInit must only be called once');
+            }
+            init = {
+              deps: regInit.deps,
+              func: regInit.init,
+            };
+          },
+        });
+
+        if (!init) {
+          throw new Error(
+            `registerInit was not called by register in ${c.moduleId} module for ${c.pluginId}`,
+          );
+        }
+
+        registrations = [
+          {
+            type: 'module',
+            pluginId: c.pluginId,
+            moduleId: c.moduleId,
+            init,
+          },
+        ];
+        return registrations;
+      },
     };
-  }
-  return () => ({
-    id: `${config.pluginId}.${config.moduleId}`,
-    register(register: BackendRegistrationPoints) {
-      return config.register({
-        registerInit: register.registerInit.bind(register),
-      });
-    },
-  });
+  };
 }