+++ packages/backend-test-utils/src/next/services/mockServices.ts
@@ -25,32 +25,36 @@ import {
   TokenManagerService,
 } from '@backstage/backend-plugin-api';
 import {
-  cacheFactory,
-  databaseFactory,
-  httpRouterFactory,
-  lifecycleFactory,
-  loggerFactory,
-  permissionsFactory,
-  rootLifecycleFactory,
-  schedulerFactory,
-  urlReaderFactory,
+  cacheServiceFactory,
+  databaseServiceFactory,
+  httpRouterServiceFactory,
+  lifecycleServiceFactory,
+  loggerServiceFactory,
+  permissionsServiceFactory,
+  rootLifecycleServiceFactory,
+  schedulerServiceFactory,
+  urlReaderServiceFactory,
 } from '@backstage/backend-app-api';
 import { ConfigReader } from '@backstage/config';
 import { JsonObject } from '@backstage/types';
 import { MockIdentityService } from './MockIdentityService';
 import { MockRootLoggerService } from './MockRootLoggerService';
 
-function simpleFactory<TService, TOptions extends [options?: object] = []>(
-  ref: ServiceRef<TService>,
+function simpleFactory<
+  TService,
+  TScope extends 'root' | 'plugin',
+  TOptions extends [options?: object] = [],
+>(
+  ref: ServiceRef<TService, TScope>,
   factory: (...options: TOptions) => TService,
-): (...options: TOptions) => ServiceFactory<TService> {
+): (...options: TOptions) => ServiceFactory<TService, TScope> {
   return createServiceFactory((options: unknown) => ({
     service: ref as ServiceRef<TService, any>,
     deps: {},
     async factory() {
       return (factory as any)(options);
     },
-  }));
+  })) as (...options: TOptions) => ServiceFactory<TService, any>;
 }
 
 /**
@@ -67,13 +71,11 @@ export namespace mockServices {
   }
 
   export function rootLogger(options?: rootLogger.Options): LoggerService {
-    return new MockRootLoggerService(options?.levels ?? false, {});
+    return MockRootLoggerService.create(options);
   }
   export namespace rootLogger {
     export type Options = {
-      levels:
-        | boolean
-        | { error: boolean; warn: boolean; info: boolean; debug: boolean };
+      level?: 'none' | 'error' | 'warn' | 'info' | 'debug';
     };
 
     export const factory = simpleFactory(coreServices.rootLogger, rootLogger);
@@ -109,30 +111,30 @@ export namespace mockServices {
   //               some may need a bit more refactoring for it to be simpler to
   //               re-implement functioning mock versions here.
   export namespace cache {
-    export const factory = cacheFactory;
+    export const factory = cacheServiceFactory;
   }
   export namespace database {
-    export const factory = databaseFactory;
+    export const factory = databaseServiceFactory;
   }
   export namespace httpRouter {
-    export const factory = httpRouterFactory;
+    export const factory = httpRouterServiceFactory;
   }
   export namespace lifecycle {
-    export const factory = lifecycleFactory;
+    export const factory = lifecycleServiceFactory;
   }
   export namespace logger {
-    export const factory = loggerFactory;
+    export const factory = loggerServiceFactory;
   }
   export namespace permissions {
-    export const factory = permissionsFactory;
+    export const factory = permissionsServiceFactory;
   }
   export namespace rootLifecycle {
-    export const factory = rootLifecycleFactory;
+    export const factory = rootLifecycleServiceFactory;
   }
   export namespace scheduler {
-    export const factory = schedulerFactory;
+    export const factory = schedulerServiceFactory;
   }
   export namespace urlReader {
-    export const factory = urlReaderFactory;
+    export const factory = urlReaderServiceFactory;
   }
 }