+++ plugins/github-pull-requests-board/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-github-pull-requests-board
 
+## 0.1.9-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.2.0-next.1
+  - @backstage/core-components@0.12.4-next.1
+  - @backstage/core-plugin-api@1.3.0
+  - @backstage/integration@1.4.2
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog-react@1.3.0-next.2
+
 ## 0.1.9-next.1
 
 ### Patch Changes