+++ plugins/jenkins/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-jenkins
 
+## 0.7.14-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.2.0-next.1
+  - @backstage/core-components@0.12.4-next.1
+  - @backstage/core-plugin-api@1.3.0
+  - @backstage/errors@1.1.4
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog-react@1.3.0-next.2
+  - @backstage/plugin-jenkins-common@0.1.13-next.1
+
 ## 0.7.14-next.1
 
 ### Patch Changes