+++ plugins/user-settings/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-user-settings
 
+## 0.6.3-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.4-next.1
+  - @backstage/core-app-api@1.4.1-next.0
+  - @backstage/core-plugin-api@1.3.0
+  - @backstage/errors@1.1.4
+  - @backstage/theme@0.2.16
+  - @backstage/types@1.0.2
+  - @backstage/plugin-catalog-react@1.3.0-next.2
+
 ## 0.6.3-next.1
 
 ### Patch Changes